# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rese",
    "description": "Rese API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server (php artisan serve)"
    },
    {
      "url": "http://rese-backend.test",
      "description": "Local server (Herd)"
    }
  ],
  "paths": {
...
    "/customers/{customer}": {
      "get": {
        "tags": ["Customer"],
        "summary": "顧客情報取得",
        "description": "セッション中の顧客情報を取得する",
        "operationId": "get-customer",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/show-customer-200"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          }
        }
      }
    },
    "/customers/{customer}/favorites": {
      "get": {
        "tags": ["Favorite"],
        "summary": "マイページでのお気に入り一覧取得機能",
        "description": "セッション中の顧客がお気に入り登録している飲食店の一覧を取得する",
        "operationId": "get-customer-favorites",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          },
          {
            "$ref": "#/components/parameters/page-query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/get-customer-favorites-200"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          }
        }
      }
    },
    "/customers/{customer}/reservations": {
      "get": {
        "tags": ["Reservation"],
        "summary": "マイページでの予約一覧取得機能",
        "description": "セッション中の顧客が行っている予約の一覧を取得する",
        "operationId": "get-customer-reservations",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/get-customer-reservations-200"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          }
        }
      }
    },
    "/customers/{customer}/reservations/{reservation}": {
      "put": {
        "tags": ["Reservation"],
        "summary": "マイページでの予約変更機能",
        "description": "セッション中の顧客が行っている指定の予約を変更する",
        "operationId": "put-customer-reservation",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          },
          {
            "$ref": "#/components/parameters/reservation-id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/put-customer-reservation"
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/no-content"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          },
          "422": {
            "$ref": "#/components/responses/put-customer-reservation-422"
          }
        }
      },
      "delete": {
        "tags": ["Reservation"],
        "summary": "マイページでの予約取り消し機能",
        "description": "セッション中の顧客が指定の飲食店で行っている指定の予約を取り消す",
        "operationId": "delete-customer-reservations",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          },
          {
            "$ref": "#/components/parameters/reservation-id"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/no-content"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          }
        }
      }
    },
    "/customers/{customer}/shops/{shop}/favorite": {
      "post": {
        "tags": ["Favorite"],
        "summary": "お気に入り登録",
        "description": "セッション中の顧客が指定の飲食店をお気に入り登録する",
        "operationId": "post-customer-shop-favorite",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          },
          {
            "$ref": "#/components/parameters/shop-id"
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/components/responses/created"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          },
          "422": {
            "$ref": "#/components/responses/unprocessable-entity"
          }
        }
      },
      "delete": {
        "tags": ["Favorite"],
        "summary": "お気に入り解除",
        "description": "セッション中の顧客が指定の飲食店のお気に入りを解除する",
        "operationId": "delete-customer-shop-favorite",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          },
          {
            "$ref": "#/components/parameters/shop-id"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/no-content"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          },
          "422": {
            "$ref": "#/components/responses/unprocessable-entity"
          }
        }
      }
    },
    "/customers/{customer}/shops/{shop}/reservations": {
      "get": {
        "tags": ["Reservation"],
        "summary": "飲食店詳細ページでの予約一覧取得機能",
        "description": "セッション中の顧客が指定の飲食店で行っている予約を一覧取得する",
        "operationId": "get-customer-shop-reservations",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          },
          {
            "$ref": "#/components/parameters/shop-id"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/get-customer-shop-reservations-200"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          }
        }
      },
      "post": {
        "tags": ["Reservation"],
        "summary": "飲食店詳細ページでの予約追加機能",
        "description": "セッション中の顧客が指定の飲食店で予約を追加する",
        "operationId": "post-customer-shop-reservations",
        "parameters": [
          {
            "$ref": "#/components/parameters/customer-id"
          },
          {
            "$ref": "#/components/parameters/shop-id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/post-customer-shop-reservations"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/post-customer-shop-reservations-201"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/forbidden"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          },
          "422": {
            "$ref": "#/components/responses/post-customer-shop-reservations-422"
          }
        }
      }
    },
...
    "/shops": {
      "get": {
        "tags": ["Shop"],
        "summary": "飲食店一覧取得",
        "description": "飲食店一覧を取得する",
        "operationId": "get-shops",
        "parameters": [
          {
            "$ref": "#/components/parameters/area-query"
          },
          {
            "$ref": "#/components/parameters/genre-query"
          },
          {
            "$ref": "#/components/parameters/search-query"
          },
          {
            "$ref": "#/components/parameters/page-query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/get-shops-200"
          }
        }
      }
    },
    "/shops/{shop}": {
      "get": {
        "tags": ["Shop"],
        "summary": "飲食店情報取得",
        "description": "飲食店情報を個別に取得する",
        "operationId": "get-shop",
        "parameters": [
          {
            "$ref": "#/components/parameters/shop-id"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/get-shop-200"
          },
          "404": {
            "$ref": "#/components/responses/not-found"
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "顧客登録",
        "description": "顧客を新規登録する",
        "operationId": "post-auth-register",
        "requestBody": {
          "$ref": "#/components/requestBodies/post-auth-register"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/created"
          },
          "204": {
            "$ref": "#/components/responses/no-content"
          },
          "422": {
            "$ref": "#/components/responses/post-auth-register-422"
          }
        }
      }
    },
    "/auth/email/verification-notification": {
      "post": {
        "tags": ["Auth"],
        "summary": "確認メール送信",
        "description": "メールアドレス確認通知を登録メールアドレスに送信する",
        "operationId": "post-auth-email-verification-notification",
        "responses": {
          "204": {
            "$ref": "#/components/responses/no-content"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "ログイン",
        "description": "顧客のログイン処理を行う",
        "operationId": "post-auth-login",
        "requestBody": {
          "$ref": "#/components/requestBodies/post-auth-login"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ok"
          },
          "204": {
            "$ref": "#/components/responses/no-content"
          },
          "422": {
            "$ref": "#/components/responses/post-auth-login-422"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "show-customer-200": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "required": ["name"],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "テストユーザー"
                }
              },
              "type": "object"
            }
          }
        }
      },
...
      "created": {
        "description": "リクエストが成功しリソースが作成された"
      },
      "no-content": {
        "description": "リクエストが成功しリソースが存在しない"
      },
...
      "unauthorized": {
        "description": "認証エラー"
      },
      "forbidden": {
        "description": "認可エラー"
      },
      "not-found": {
        "description": "指定されたリソースが存在しない"
      },
...
      "unprocessable-entity": {
        "description": "リクエストが正しくない"
      },
...
      "post-auth-register-422": {
        "description": "バリデーションエラーまたはメールアドレスが登録済み",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/register-error"
            },
            "examples": {
              "required": {
                "$ref": "#/components/examples/post-auth-register-422-required"
              },
              "invalid-email": {
                "$ref": "#/components/examples/post-auth-register-422-email"
              },
              "password": {
                "$ref": "#/components/examples/post-auth-register-422-password"
              }
            }
          }
        }
      },
...
      "get-shops-200": {
        "description": "飲食店一覧取得成功",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/pagination"
                },
                {
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/shop-data"
                      }
                    }
                  },
                  "type": "object"
                }
              ]
            },
            "examples": {
              "first-page": {
                "$ref": "#/components/examples/get-shops-200-first-page"
              },
              "middle-page": {
                "$ref": "#/components/examples/get-shops-200-middle-page"
              },
              "last-page": {
                "$ref": "#/components/examples/get-shops-200-last-page"
              },
              "empty": {
                "$ref": "#/components/examples/get-shops-200-empty"
              }
            }
          }
        }
      },
      "get-shop-200": {
        "description": "飲食店情報取得成功",
        "content": {
          "application/json": {
            "schema": {
              "required": ["data"],
              "properties": {
                "data": {
                  "$ref": "#/components/schemas/shop-data"
                }
              },
              "type": "object"
            },
            "examples": {
              "example": {
                "$ref": "#/components/examples/get-shop-200"
              }
            }
          }
        }
      },
      "get-customer-favorites-200": {
        "description": "マイページのお気に入り一覧取得成功",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/pagination"
                },
                {
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/shop-data"
                      }
                    }
                  },
                  "type": "object"
                }
              ]
            },
            "examples": {
              "example": {
                "$ref": "#/components/examples/get-customer-favorites-200"
              }
            }
          }
        }
      },
      "get-customer-reservations-200": {
        "description": "マイページの予約一覧取得成功",
        "content": {
          "application/json": {
            "schema": {
              "required": ["data"],
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/reservation-data"
                  }
                }
              },
              "type": "object"
            }
          }
        }
      },
      "get-customer-shop-reservations-200": {
        "description": "飲食店詳細ページの予約一覧取得成功",
        "content": {
          "application/json": {
            "schema": {
              "required": ["reservations"],
              "properties": {
                "reservations": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/reservation-data"
                  }
                }
              },
              "type": "object"
            },
            "examples": {
              "example": {
                "$ref": "#/components/examples/get-customer-shop-reservations-200-example"
              },
              "empty": {
                "$ref": "#/components/examples/get-customer-shop-reservations-200-empty"
              }
            }
          }
        }
      },
      "post-customer-shop-reservations-201": {
        "description": "飲食店詳細ページの予約追加成功",
        "content": {
          "application/json": {
            "schema": {
              "required": ["reservation"],
              "properties": {
                "reservation": {
                  "$ref": "#/components/schemas/reservation-data"
                }
              },
              "type": "object"
            },
            "examples": {
              "example": {
                "$ref": "#/components/examples/post-customer-shop-reservations-201"
              }
            }
          }
        }
      },
      "post-customer-shop-reservations-422": {
        "description": "飲食店詳細ページの予約追加のバリデーションエラー",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/reservation-error"
            },
            "examples": {
              "required": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-required"
              },
              "date": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-reserved_at"
              },
              "past": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-past"
              },
              "number_of_guests": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-number_of_guests"
              }
            }
          }
        }
      },
      "put-customer-reservation-422": {
        "description": "マイページの予約変更のバリデーションエラー",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/reservation-error"
            },
            "examples": {
              "required": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-required"
              },
              "date": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-reserved_at"
              },
              "past": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-past"
              },
              "number_of_guests": {
                "$ref": "#/components/examples/post-customer-shop-reservations-422-number_of_guests"
              }
            }
          }
        }
      }
    },
    "parameters": {
...
      "customer-id": {
        "name": "customer",
        "in": "path",
        "description": "顧客ID",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "example": 1
        }
      },
      "shop-id": {
        "name": "shop",
        "in": "path",
        "description": "飲食店ID",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "example": 1
        }
      },
      "reservation-id": {
        "name": "reservation",
        "in": "path",
        "description": "予約ID",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "example": 1
        }
      },
      "area-query": {
        "name": "area",
        "in": "query",
        "description": "エリアID",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "genre-query": {
        "name": "genre",
        "in": "query",
        "description": "ジャンルID",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "search-query": {
        "name": "search",
        "in": "query",
        "description": "店名検索キーワード",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "page-query": {
        "name": "page",
        "in": "query",
        "description": "ページ番号",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "requestBodies": {
      "post-auth-register": {
        "description": "顧客登録リクエスト",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "required": ["name", "email", "password"],
              "properties": {
                "name": {
                  "description": "顧客名",
                  "type": "string",
                  "maxLength": 100,
                  "minLength": 1,
                  "example": "テストユーザー"
                },
                "email": {
                  "description": "メールアドレス",
                  "type": "string",
                  "format": "email",
                  "example": "test@example.com"
                },
                "password": {
                  "description": "パスワード",
                  "type": "string",
                  "maxLength": 100,
                  "minLength": 8,
                  "example": "password"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "post-auth-login": {
        "description": "ログインリクエスト",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "required": ["email", "password"],
              "properties": {
                "email": {
                  "description": "メールアドレス",
                  "type": "string",
                  "format": "email",
                  "example": "test@example.com"
                },
                "password": {
                  "description": "パスワード",
                  "type": "string",
                  "maxLength": 100,
                  "minLength": 8,
                  "example": "password"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "post-customer-shop-reservations": {
        "description": "予約追加リクエスト",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "required": ["reserved_at", "number_of_guests"],
              "properties": {
                "reserved_at": {
                  "description": "予約日時",
                  "type": "string",
                  "format": "date-time",
                  "example": "2021-12-31T23:59:59+09:00"
                },
                "number_of_guests": {
                  "description": "予約人数",
                  "type": "integer",
                  "minimum": 1,
                  "example": 2
                }
              },
              "type": "object"
            }
          }
        }
      },
      "put-customer-reservation": {
        "description": "予約変更リクエスト",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "required": ["reserved_at", "number_of_guests"],
              "properties": {
                "reserved_at": {
                  "description": "予約日時",
                  "type": "string",
                  "format": "date-time",
                  "example": "2021-12-31T23:59:59+09:00"
                },
                "number_of_guests": {
                  "description": "予約人数",
                  "type": "integer",
                  "minimum": 1,
                  "example": 2
                }
              },
              "type": "object"
            }
          }
        }
      }
    }
  },
...
}